Created LocationGeocoder class, updated profileInfoActivity and EventAttendeesArrayAdapter to use new accessUserProfileImage method
